delete-resorthome

+++ src/pages/attence/resorthome.vue
@@ -220,7 +220,14 @@ export default {
         },
         delete: scope => {
           const id = scope.row.waid;
-          this.deleteAttendance(id);
+
+          this.$confirm("此操作永久删除该班级, 是否继续?", "提示", {
+            confirmButtonText: "确定",
+            cancelButtonText: "取消",
+            type: "warning"
+          }).then(() => {
+            this.deleteAttendance(id);
+          });
         }
       };
     },
